chore(settings): remove the Customization tab
Customization (internal tab id `shell`) is now fully managed locally in the
server, so the desktop tab is dead weight. Deletes the tab view and its four
registration points.

The Windows/Linux "Display menu bar" toggle moves to Appearance > Window: it
drives `applicationMenuVisible` over IPC to the native Electron menu and had no
other UI surface, so deleting it outright would have orphaned the feature.

`shell/shell-config.tsx` and `ui-state-store.ts` stay untouched - they are read
by the status bar, sidebar, notifications, starter cards, and welcome page.
